Time: 76 ms (28.33%), Space: 33.2 MB (20.39%) - LeetHub

+++ 0264-ugly-number-ii/0264-ugly-number-ii.cpp
@@ -8,7 +8,7 @@ public:
         pq.push(1);
         seen.insert(1);
 
-        int curr = 1;
+        long curr = 1;
 
         while(n--)
         {
